shutdown on bad wifi config

+++ home.admin/_bootstrap.sh
@@ -160,6 +160,14 @@ if [ "${flagExists}" == "1" ]; then
   # set wifi
   echo "Setting Wifi SSID(${ssid}) Password(${password})" >> ${logFile}
   /home/admin/config.scripts/internet.wifi.sh on ${ssid} ${password} >> ${logFile}
+  if [ $? -gt 0 ]; then
+    echo "Display failure for 1min on on LCD display" >> ${logFile}
+    /home/admin/_cache.sh set state "errorWIFI"
+    /home/admin/_cache.sh set message "shutting down - edit or remove file"
+    sleep 60
+    sudo shutdown now
+    exit 1
+  fi
 
   # remove flag
   sudo rm /boot/firmware/wifi
